Corrected notes, added POJO classes.

+++ app/src/main/java/ru/fadedfog/tombs/asset/geometry/Point.java
@@ -1,31 +1,31 @@
 package ru.fadedfog.tombs.asset.geometry;
 
 public class Point {
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	public Point() {
 		
 	}
 	
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 	
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 	
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
